Guard against bad/null inputs in the Array API, and change the signature to prevent negative counts and sizes (which would make no sense).

+++ include/pool.h
@@ -27,7 +27,7 @@
 /* Memory allocation/anti-leak system.  Yes, this *IS* stolen from Apache
  * also.  What can I say?  It makes sense, and it's safe (more overhead
  * though)
- * $Id: pool.h,v 1.21 2008-02-11 04:37:49 castaglia Exp $
+ * $Id: pool.h,v 1.22 2008-02-13 07:24:36 castaglia Exp $
  */
 
 #ifndef PR_POOL_H
@@ -66,7 +66,7 @@ typedef struct {
   void *elts;
 } array_header;
 
-array_header *make_array(pool *, int, int);
+array_header *make_array(pool *, unsigned int, size_t);
 void *push_array(array_header *);
 void array_cat(array_header *, const array_header *);
 array_header *append_arrays(pool *, const array_header *, const array_header *);
